Break codebase into separate pieces
In particular, breaks the Contract class into its own file.

This has the benefit of being more modular, allowing subclassing for external modules, as well as saving memory that was previously wasted re-initializing the class on each instance creation.

Fixes #3.
