+++ src/has-tx-object.js
@@ -1,0 +1,16 @@
+const arrayContainsArray = require('ethjs-util').arrayContainsArray;
+
+function hasTransactionObject(args) {
+  const txObjectProperties = ['from', 'to', 'data', 'value', 'gasPrice', 'gas'];
+  if (typeof args === 'object' && Array.isArray(args) === true && args.length > 0) {
+    if (typeof args[args.length - 1] === 'object'
+      && (Object.keys(args[args.length - 1]).length === 0
+      || arrayContainsArray(Object.keys(args[args.length - 1]), txObjectProperties, true))) {
+      return true;
+    }
+  }
+
+  return false;
+}
+
+module.exports = hasTransactionObject;
